image_retriever/llava/requirements.txt
transformers==4.36.2
gradio
torch
Pillow